provide a template file for the environment settings and ensure it exists when setup.sh is run

+++ demo_scripts/build_2016/setup.sh
@@ -1,3 +1,14 @@
+# Ensure Azure storage confgiuration is completed
+if [ ! -f env.conf ]
+then
+    cp env.conf.tmpl env.conf
+    echo "=========================== NOTICE ==============================="
+    echo "You will need to set the Azure storage account details in env.conf"
+    echo "Once configured rerun setup.sh"
+    echo "=================================================================="
+    exit 1
+fi
+
 # Stop and remove any containers related to this demo
 docker-compose stop
 docker-compose rm -f
@@ -7,3 +18,4 @@ docker run --env-file env.conf rgardler/acs-logging-test-cli deleteQueue
 docker run --env-file env.conf rgardler/acs-logging-test-cli deleteTable
 docker run --env-file env.conf rgardler/acs-logging-test-cli createQueue
 docker run --env-file env.conf rgardler/acs-logging-test-cli createTable
+
